:sparkles: (infra): add infrastructure layer
